// packages/core/src/db/schema.ts
export interface FinancialGoalTable {
  id: string;
  userId: string;
  name: string;
  /** paise */
  targetAmount: number;
  /** paise */
  currentAmount: number;
  targetDate: string | null;
  description: string | null;
  color: string;
  createdAt: string;
  updatedAt: string;
}